+++ token.go
@@ -12,7 +12,7 @@ type Token struct {
 	text []Text
 
 	// 分词在语料库中的词频
-	frequency int
+	frequency float64
 
 	// log2(总词频/该分词词频)，这相当于 log2(1/p(分词))，用作动态规划中
 	// 该分词的路径长度。求解 prod(p(分词)) 的最大值相当于求解
@@ -32,7 +32,7 @@ func (token *Token) Text() string {
 }
 
 // Frequency 返回分词在语料库中的词频
-func (token *Token) Frequency() int {
+func (token *Token) Frequency() float64 {
 	return token.frequency
 }
